Implement register endpoint
A lot of reusable plumbing was added to aid the implementation, hence the large
diff. In particular, the filesystem generator was imported from another project
(eventually to be extracted to its own repository), snowflake ID generation was
also imported from another project (eventually to be published on crates.io),
and many dependencies were introduced.