backend: Fix startup crash in Docker container
The Cargo toolchain on the latest versions of Alpine Linux has started
dynamically linking to gcc alongside musl when the crt-static feature is
disabled (this is nonsensical and probably a bug). Since libgcc is not
available in the runtime container, the server crashes on startup.

As a workaround, install libgcc at runtime, as this is the easiest step
forward. Since we only use containers for testing, the mildly increased
image size should not be a problem.

Alternative solutions are to use a non-musl image (e.g. Ubuntu), or to
fix the libpq link errors when crt-static is enabled.